Added Mac-specific utilities and started the build of Application. BSD users: If you are using KDE as your desktop, you have X11. startx and run shell script LINUX users: If you are using KDE as your desktop, you have X11. visit /etc/X11 to configure, then run startx. Windows users: double click the jar file. Or run ServerMain and ClientMain. MacOSX users: You can run the windows jars, or just use the application stub. run cmake to build the app from source.
Note: JDK9+ deprecates the sun.misc.Unsafe class. It is provided for convenience in a jar file. (rt.jar also happens to contain the mac libs, so If you are on Windows you won't get any compiler errors if you try to run the mac main. You WILL get runtime errors as JNI looks for the header files that don't exist on your system. You can sort of do it with cygwin, but just run the jar.)