ALPHA RELEASE
Recompiled jars to fit new changes. Please verify that they work on other platforms.

This should be the final commit to version ALPHA.

Native code implementations will be skipped in favor of a wrapper of Unsafe from sun.misc package. If you have JDK 9+ please change the declarations of sun.misc.Unsafe to jdk.internal.Unsafe to retain functionality. I am working on a wrapper for it for use on all platforms. Ideally, we will be able to use it for persistent memory.

-- Michael Ferolito --