Add missing Linux jars
Mesa setup is still broken; that's an upstream problem. Launch eclipse
and/or the program with switcherooctl launch to use the dedicated gpu
instead. This bypasses Mesa and it works.

It also will not work on recent macs because of Apple's decision to
deprecate (and remove) OpenGL in favor of its own proprietary Metal API.
On a Mac? glhf